Restore the lemma test suite on the integers frontend
The cp526/integers frontend rejects the bitvector-typed CN annotations
of the legacy fixtures (i32/u64 types, suffixed literals), which is why
upstream disabled the Rocq workflow. Convert the fixtures to integer
mode (types to integer, drop literal suffixes, drop the identity casts in
struct_test) and rewrite the five handwritten proofs that depended on the
former Ptr := Z heap for the VIP byte model: list_segment, array_combine,
arrays_inductive, arrays_testing and pop_queue. All nine cases of
tests/run-cn-lemmas.sh pass and no proof in the suite uses Admitted
(upstream's arrays_testing and arrays_inductive proofs did).

One fixture lemma changes: arrays/inductive array_lemma now requires
!is_null(p). Without it the statement has a countermodel in the VIP
model, because for p = NULL and n = 0 the second Array lives at
array_shift(NULL, 0), a non-null pointer with the null selectors'
provenance, which cannot be re-rooted at NULL.

+++ tests/rocq_lemmas/cases/struct_test/resource_lemma.h
@@ -2,8 +2,8 @@
 lemma resource_lemma (pointer p)
   requires 
       take v1 = Owned<int>(p);
-      (i32) v1 == 0i32;
+      v1 == 0;
   ensures 
       take v2 = Owned<int>(p);
-      (i32) v2 == 0i32;
+      v2 == 0;
 @*/
